+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Support VSCode editor
+
 ## [0.3.0] - 2021-10-17
 
 - Add cmd/ifind for grepping and opening results